MSVC++ example build fixes.
git-svn-id: http://svn.code.sf.net/p/qsampler/code_attic/liblscp@335 3e92087d-0f25-4bc2-b788-d05a45ecc5c1

+++ examples/parser.c
@@ -66,7 +66,7 @@ float lscp_parser_nextnum ( lscp_parser_t *pParser )
     float ret = 0;
 
     if (pParser->pszToken) {
-        ret = atof(pParser->pszToken);
+        ret = (float) atof(pParser->pszToken);
         lscp_parser_next(pParser);
     }
